MSVC warnings

+++ include/fplus/container_common.h
@@ -800,7 +800,7 @@ ContainerOut generate_range_step
         (const T start, const T end, const T step)
 {
     ContainerOut result;
-    std::size_t size = (end - start) / step;
+    std::size_t size = static_cast<std::size_t>((end - start) / step);
     prepare_container(result, size);
     auto it = get_back_inserter<ContainerOut>(result);
     for (T x = start; x < end; x+=step)
